Implement Server Layer API for data requests
This commit adds custom request handlers for the various CRUD operations
involved with manipulating heroes. The app now queries the api/ url to
perform the operations so the server can respond appropriately. This
moves the memory persistence from in-memory of the browser to in-memory
of the server (e.g. data is now persisted across page refreshes).

This is the next step in moving memory persistence to a database level.

Ticket/Issue: N/A