Add MongoDb Support
This commit adds mongoldb support with a new version of the express
server. This allows a connection to a mongodb database and performs all
the same api requests by querying the db for data. This provides a
memory persistence with a separate instance of the database instead of
the server itself, allowing the server to restarted without having
the data change.

Ticket/Issue: N/A